+++ js/data.js
@@ -1,0 +1,26 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'Имена авторов также должны быть случайными. Набор имён для комментаторов составьте сами. Подставляйте случайное имя в поле name.'
+];
+
+const DESCRIPTION = [
+  'котики',
+  'собачки',
+  'природа'
+];
+
+const NAMES = [
+  'Вася',
+  'Петя',
+  'Коля',
+  'Оля',
+  'Маша',
+  'Даша'
+];
+
+export {MESSAGES, DESCRIPTION, NAMES};
