+++ js/util.js
@@ -1,0 +1,11 @@
+const getRandomPositiveInteger = ( numberA, numberB) => {
+  const lower = Math.ceil(Math.min(Math.abs(numberA), Math.abs(numberB)));
+  const upper = Math.floor(Math.max(Math.abs(numberA), Math.abs(numberB)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomPositiveInteger, getRandomArrayElement};
+
